MAP2_S01: func_800CED88

+++ include/maps/shared.h
@@ -2574,7 +2574,7 @@ void sharedFunc_800D9064_0_s00(s_SubCharacter* chara);
 
 void sharedFunc_800D9078_0_s00(s_SubCharacter* chara);
 
-s16 Anim_StartKeyframeIdxGet(s_SubCharacter* chara);
+s32 Anim_StartKeyframeIdxGet(s_SubCharacter* chara);
 
 /** Humanoid init function? */
 void sharedFunc_800D923C_0_s00(s_SubCharacter* chara);
